Reorg and add low level funcs for objects
